pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fl_ids"
version = "1.0.0"
description = "Federated Learning for Intrusion Detection System"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.25.0",
    "tensorflow-cpu>=2.9.1, != 2.11.1 ; platform_machine == \"x86_64\"",
    "pandas",
    "scikit-learn",
    "wandb==0.17.8",
]
[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "fl_ids.server_app:app"
clientapp = "fl_ids.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 10
local-epochs = 1
batch-size = 64
min-available-clients = 3
min-fit-clients = 3
min-evaluate-clients = 3
fraction-train = 1.0
fraction-evaluate = 1.0

[tool.flwr.federations]
default = "local-sim"

[tool.flwr.federations.local-sim]
options.num-supernodes = 50
options.backend.client-resources.num-cpus = 2 # each ClientApp assumes to use 2CPUs
options.backend.client-resources.num-gpus = 0.0 # ratio of VRAM a ClientApp has access to

[tool.flwr.federations.local-sim-gpu]
options.num-supernodes = 50
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0.25